src/Backend/requirements.txt
boto3
botocore
fastapi
mlflow[extras]
numpy
onnxruntime_gpu
opencv-python-headless
Pillow
uvicorn
python-multipart
